fix: detect support for workers
Some contexts (e.g. service workers) do not support web workers, so
when they are not supported, fall back to running OPFS operations
asynchronously in the main thread.
