+++ .gitignore
@@ -2,3 +2,4 @@ node_modules/
 test-results/
 dist/
 dist-worker/
+.vscode
