+++ src/opfs-worker.ts
@@ -916,4 +916,4 @@ async function Se(e, t) {
 		result: null
 	});
 }
-//#endregion`;
+//#endregion`
